requirements.txt
annotated-types==0.5.0
blis==0.7.11
catalogue==2.0.10
certifi==2022.12.7
charset-normalizer==3.3.2
ChatterBot==1.0.0
chatterbot-corpus==1.2.0
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
cymem==2.0.8
idna==3.10
importlib-metadata==6.7.0
Jinja2==3.1.4
joblib==1.3.2
langcodes==3.3.0
markdown-it-py==2.2.0
MarkupSafe==2.1.5
mathparse==0.1.2
mdurl==0.1.2
murmurhash==1.0.10
nltk==3.8.1
numpy==1.21.6
packaging==24.0
Pint==0.18
pip==22.3.1
preshed==3.0.9
pydantic==1.7.4
pydantic_core==2.14.6
Pygments==2.17.2
pymongo==3.13.0
python-dateutil==2.7.5
pytz==2024.2
PyYAML==3.13
regex==2024.4.16
requests==2.31.0
rich==13.8.1
setuptools==65.6.3
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.2.19
srsly==2.4.8
thinc==8.2.5
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.7.1
urllib3==2.0.7
wasabi==0.10.1
weasel==0.4.1
wheel==0.38.4
wincertstore==0.2
wrapt==1.16.0
zipp==3.15.0
